backend/requirements.txt
fastapi
uvicorn
python-dotenv
ollama
google-genai
